Update initramfs with new uClibc
For the kernel changes to use generic stat.h, a synchronized change to
uClibc is required.  This changeset updates the initramfs with a new
busybox that has been built against the updated uClibc.

This patch and the previous one go together as the initramfs and kernel
will not run properly without these two changes happening concurrently.