Remove vfork syscall
"vfork is deprecated"... at least it shouldn't be necessary on systems with
an MMU as fork does what vfork does just as well.

For some reason vfork in busybox telnetd works strangely and causes the
parent process' stack to be unwound.  Fork, on the other hand, works and
uClibc is happy to map vfork to fork for us if the system call is absent and
we have an MMU.

Of course, if we don't have an MMU, it's another story, but is this really
even relevant for OR32?  In theory, yes, but in practice...?