Clear SPR_TTMR_IP after servicing interrupt
The IP bit of the TTMR bit must be cleared after servicing the timer interrupt,
otherwise no further interrupts will be issued.  This was 'incidentally' done
before by rewriting the entire TTMR register in the timer_interrupt function,
but this should be done more explicitly at the tail of the interrupt exception
handler.