Pull in updated spr_defs.h
This updated version is the "golden version" which is maintained in the
or1ksim project.  Most of the churn in this patch is due to someone messing
up the whitespace... no worries, we aim to keep this in sync with the
or1ksim version from now on so we'll accept their whitespace!

Primary changes here are:
- rename SPR_PC to SPR_NPC
- rename SPR_ICCFGR_BS to SPR_ICCFGR_CBS
- rename SPR_DCCFGR_BS to SPR_DCCFGR_CBS
- major cleanup of SPR_UPR; it actually matches spec now!