Remove symbol prefix for updated GCC
The OR32 GCC compiler has been changed to stop prepending "_" to symbol names
in order to be in more in line with what other architectures do.  This
patch does the necessary modifications to account for that change.

Note that this patch breaks compilation with older OR32 compilers.  This is
a conscious decision that has been made.