Switch to generic IRQ handling
This is a big patch that switches the IRQ handling to using as much generic
code as possible.  The IRQ controller is rewritten to use struct irq_chip
and the IRQ handling is turned over to the generic code.

There are a fair number of cleanups associated with this code that were
necessary to allow compilation.

Contrary to the spec, it seems that a 0 has to be written to PICSR
to clear pending interrupt.  The spec says a 1 should be written.