ethoc: rework interrupt handling
The old interrupt handling was incorrect in that it did not account for the
fact that the interrupt source bits get set irregardless of whether or not
their corresponding mask is set.  This patch fixes that by masking off the
source bits for masked interrupts.  Furthermore, the handling of
transmission events is moved to the NAPI polling handler alongside the
reception handler, thus preventing a whole bunch of interrupts during
heavy traffic.

Signed-off-by: Jonas Bonn <jonas@southpole.se>

Conflicts:

	drivers/net/ethoc.c