Add fault handlers
Add handlers for unaligned access, bus fault, and illegal instruction.  None of
these handlers attempts to handle the fault in any way other than signalling
it to the user process.

The unaligned access handler may want to be improved in the future to
transparently handle the access on behalf of the process.

Signed-off-by: Jonas Bonn <jonas@southpole.se>