Pass valid bitmap page to init_bootmem
The first argument to init_bootmem needs to be the page where the bitmap
should go.  Passing page 0 here was a particularly bad idea as it would
lead to the exception handlers being overwritten by the bitmap.  This
patch changes the parameter to pass the first page after the kernel
instead, as it should be.

Debugged by setting page 0 read-only and getting the PC from the
exception.  Page 0 should really always have been read-only, but there's
some self-modifying code lying around that wants to write there, as
well as the temporary stores for the exception handler (which are done
with MMU off anyway, so they're probably OK).