Clear r0 at startup
r0 is "always zero" by convention only and there's no guarantee that
the hardware actually sets its value to zero, so do it explicitly at
startup just to be sure.  This doesn't actually cost anything as we
can sneak it into the delay slot of the jump in the reset vector where
we previously had a l.nop.