Select OF_EARLY_FLATTREE
The config option for using flattened device trees has been split up
into a generic part for general use and a part for those who need it
early on in the boot process.  As OpenRISC depends on the early usage
of the flattened tree, we need this option instead of the other one.