Rework system.h and atomic.h
This is a general rework of system.h whereby we fallback to the generic
version.  Some fixups to the handling of atomics and switch_to are
included here, too, to make the generic system.h work.

The generic system.h define nop() in a non-portable way so that needs
to be fixed long term, but for now it's ok as we don't use that function
anywhere.

It's not totally clear if we should be using the generic system.h right
off like this, but it works ok and can be reevaluated later.