Remove include of user.h from elfcore.h
As it shouldn't be necessary for new architectures to implement asm/user.h
anymore, elfcore.h should not unconditionally pull it in.  Nothing in
elfcore.h relies on struct user, in any case.

Also, fix up the comment in inotify_user.c so we don't fool ourselves into
thinking that file uses struct user, which it doesn't.

Signed-off-by: Jonas Bonn <jonas.bonn@gmail.com>