Add support for kernel modules
This also introduces a change to VMALLOC_START and _END.  Kernel modules
need to lie with a 28-bit signed offset of the kernel text so that l.j
and l.jal instructions can be used without having to introduce any
trampolines.

This also fixes a secondary bug whereby the kernel vmalloc'ed area would
collide with kmalloc'ed area when using lots (more than 256MB) of memory.