Fix read-only mapping of kernel
The main fix here is that the kernel_ro tags in the link script are
used correctly to determine the region that should be read-only mapped.

An additional fix is to flush the TLB after switching TLB-miss handlers
so that the page flags get updated as per the newly created page
tables.