Make Wishbone bus explicit
The OpenRISC SoC platform is currently only defined for a Wishbone
interconnect with 'native endian' devices, i.e. devices synthesized
together with the CPU and of the same endianess as the CPU.  This
patch introduces accessors to make that explicit in the __iomem
accessors.

With this change, we can safely say that IO access not using __iomem
accessors has undefined behaviour and that __iomem cookies can
safely be used on the OpenRISC platform.

If another interconnect is introduced that changes any of these
assumptions, then this will need to be reevaluated.