ptrace: export registers via regset only
This cleans up the ptrace interface so that registers are exported via
the PTRACE_GETREGSET mechanism.  This exports the registers in the same
format as they are exported via a coredump.

The functionality to actually export registers in a coredump is added.

Some defines needed for NOMMU architectures only (PT_TEXT_ADDR, etc.) are
removed.