openrisc: Clean up signal handling
Strange pthreads behaviour resulted in the discovery that the signal
handling for OpenRISC wasn't particularly correct.  In particular, the
syscall restart path was incorrect.

This patch is a big cleanup of the signal handling code, including making
it look more like other architectures so that it's easier for others to
follow.

Reported-by: Anders Nystrom <anders@southpole.se>
Signed-off-by: Jonas Bonn <jonas@southpole.se>