spi: litespi: fix litespi cs handling for bulk transfers
For bulk transfers it is essential to keep cs asserted for all
transfers in a block. This is done by the higher level api
trough set_cs callback.
Changes in this commit require changes in the HDL part
(https://github.com/enjoy-digital/litex) in the SPIMaster module,
namely CS register must directly drive the cs pin.